bpf: split nat.h into the loopkup part and the rest
CTLB does not need anything except the lookup part and sharing the nat.h
with the rest that is used in tc only makes it difficult to configure at
load time.
